refactor(#17): move tag editor to component
