+++ src/app/App/MainScreen/NotesOverview/NotesOverview.css
@@ -19,22 +19,3 @@
 .NotesOverview-Tags {
 	margin-top: 0.5rem;
 }
-
-.NotesOverview-TagEditor .Modal-Content {
-	display: flex;
-	flex-direction: column;
-	padding: 1.5rem 1rem;
-	min-width: 500px;
-	gap: 1rem;
-}
-
-.NotesOverview-TagEditorControls {
-	display: flex;
-	gap: 0.5rem;
-	justify-content: end;
-}
-
-.NotesOverview-TagsListInPopup .Menu-Container {
-	max-height: 300px;
-    overflow: auto;
-}
