+++ src/app/App/MainScreen/NotesOverview/TagEditor/TagEditor.css
@@ -1,0 +1,18 @@
+.TagEditor .Modal-Content {
+	display: flex;
+	flex-direction: column;
+	padding: 1.5rem 1rem;
+	min-width: 500px;
+	gap: 1rem;
+}
+
+.TagEditor-Controls {
+	display: flex;
+	gap: 0.5rem;
+	justify-content: end;
+}
+
+.TagEditor-TagsListInPopup .Menu-Container {
+	max-height: 300px;
+    overflow: auto;
+}
